move and regenerate files
