+++ generated-sources/main/java/com/shapesecurity/shift/es2017/parser/RangeCheckerReducer.java
@@ -1,4 +1,4 @@
-// Generated by reducer.js
+// Generated by rangechecker.js
 /**
  * Copyright 2018 Shape Security, Inc.
  *
@@ -23,8 +23,6 @@ import com.shapesecurity.functional.data.Maybe;
 import com.shapesecurity.functional.data.Monoid;
 import com.shapesecurity.shift.es2017.ast.*;
 import com.shapesecurity.shift.es2017.reducer.MonoidalReducer;
-import com.shapesecurity.shift.es2017.ast.Module;
-
 import javax.annotation.Nonnull;
 
 import static org.junit.Assert.assertTrue;
