+++ generated-sources/main/java/com/shapesecurity/shift/es2017/reducer/MonoidalReducer.java
@@ -1,4 +1,4 @@
-// Generated by reducer.js
+// Generated by reducer-monoidal.js
 /**
  * Copyright 2018 Shape Security, Inc.
  *
@@ -20,7 +20,6 @@ package com.shapesecurity.shift.es2017.reducer;
 
 import com.shapesecurity.functional.data.*;
 import com.shapesecurity.shift.es2017.ast.*;
-import com.shapesecurity.shift.es2017.ast.Module;
 
 import javax.annotation.Nonnull;
