+++ generated-sources/main/java/com/shapesecurity/shift/es2017/reducer/ReconstructingReducer.java
@@ -1,4 +1,4 @@
-// Generated by reducer.js
+// Generated by reducer-clone.js
 /**
  * Copyright 2018 Shape Security, Inc.
  *
@@ -21,8 +21,6 @@ package com.shapesecurity.shift.es2017.reducer;
 import com.shapesecurity.functional.data.ImmutableList;
 import com.shapesecurity.functional.data.Maybe;
 import com.shapesecurity.shift.es2017.ast.*;
-import com.shapesecurity.shift.es2017.ast.Module;
-
 import javax.annotation.Nonnull;
 
 public class ReconstructingReducer implements Reducer<Node> {
