+++ generated-sources/main/java/com/shapesecurity/shift/es2017/serialization/Deserializer.java
@@ -1,4 +1,4 @@
-// Generated by reducer.js
+// Generated by deserializer.js
 /**
  * Copyright 2018 Shape Security, Inc.
  *
@@ -16,14 +16,13 @@
  */
 
 
-package com.shapesecurity.shift.es2017.serialization;
+package com.shapesecurity.shift.serialization;
 
 import com.google.gson.*;
 
 import com.shapesecurity.functional.data.ImmutableList;
 import com.shapesecurity.functional.data.Maybe;
 import com.shapesecurity.shift.es2017.ast.*;
-import com.shapesecurity.shift.es2017.ast.Module;
 import com.shapesecurity.shift.es2017.ast.operators.BinaryOperator;
 import com.shapesecurity.shift.es2017.ast.operators.CompoundAssignmentOperator;
 import com.shapesecurity.shift.es2017.ast.operators.UnaryOperator;
@@ -38,16 +37,11 @@ public class Deserializer {
     protected Deserializer() {}
 
     public static Node deserialize(String toDeserialize) throws JSONException, IllegalAccessException, InstantiationException, ClassNotFoundException, NoSuchMethodException {
-        try {
-            JsonElement json = new JsonParser().parse(toDeserialize);
-            return new Deserializer().deserializeNode(json);
-        } catch (JsonSyntaxException e) {
-            e.printStackTrace();
-            return null;
-        }
+        JsonElement json = new JsonParser().parse(toDeserialize);
+        return deserializeNode(json);
     }
 
-    private BinaryOperator deserializeBinaryOperator(JsonElement jsonElement) {
+    private static BinaryOperator deserializeBinaryOperator(JsonElement jsonElement) {
         String operatorString = jsonElement.getAsString();
         switch (operatorString) {
             case ",":
@@ -61,7 +55,7 @@ public class Deserializer {
             case "^":
                 return BinaryOperator.BitwiseXor;
             case "&":
-                return BinaryOperator.BitwiseAnd;
+                return BinaryOperator.LogicalAnd;
             case "+":
                 return BinaryOperator.Plus;
             case "-":
@@ -98,14 +92,12 @@ public class Deserializer {
                 return BinaryOperator.Right;
             case ">>>":
                 return BinaryOperator.UnsignedRight;
-            case "**":
-                return BinaryOperator.Exp;
             default:
-                throw new RuntimeException("Unexpected binary operator in deserialization: " + operatorString);
+                return null; // should not get here
         }
     }
 
-    private CompoundAssignmentOperator deserializeCompoundAssignmentOperator(JsonElement jsonElement) {
+    private static CompoundAssignmentOperator deserializeCompoundAssignmentOperator(JsonElement jsonElement) {
         String operatorString = jsonElement.getAsString();
         switch (operatorString) {
             case "+=":
@@ -130,14 +122,12 @@ public class Deserializer {
                 return CompoundAssignmentOperator.AssignBitXor;
             case "&=":
                 return CompoundAssignmentOperator.AssignBitAnd;
-            case "**=":
-                return CompoundAssignmentOperator.AssignExp;
             default:
-                throw new RuntimeException("Unexpected compound assignment operator in deserialization: " + operatorString);
+                return null; // should not get here
         }
     }
 
-    private UnaryOperator deserializeUnaryOperator(JsonElement jsonElement) {
+    private static UnaryOperator deserializeUnaryOperator(JsonElement jsonElement) {
         String operatorString = jsonElement.getAsString();
         switch (operatorString) {
             case "+":
@@ -155,11 +145,11 @@ public class Deserializer {
             case "delete":
                 return UnaryOperator.Delete;
             default:
-                throw new RuntimeException("Unexpected unary operator in deserialization: " + operatorString);
+                return null;
         }
     }
 
-    private UpdateOperator deserializeUpdateOperator(JsonElement jsonElement) {
+    private static UpdateOperator deserializeUpdateOperator(JsonElement jsonElement) {
         String operatorString = jsonElement.getAsString();
         switch (operatorString) {
             case "++":
@@ -167,11 +157,11 @@ public class Deserializer {
             case "--":
                 return UpdateOperator.Decrement;
             default:
-                throw new RuntimeException("Unexpected update operator in deserialization: " + operatorString);
+                return null;
         }
     }
 
-    private VariableDeclarationKind deserializeVariableDeclarationKind(JsonElement jsonElement) {
+    private static VariableDeclarationKind deserializeVariableDeclarationKind(JsonElement jsonElement) {
         String kindString = jsonElement.getAsString();
         switch (kindString) {
             case "var":
@@ -181,11 +171,11 @@ public class Deserializer {
             case "let":
                 return VariableDeclarationKind.Let;
             default:
-                throw new RuntimeException("Unexpected variable declaration kind in deserialization: " + kindString);
+                return null;
         }
     }
 
-    private ImmutableList<Maybe<AssignmentTargetAssignmentTargetWithDefault>> deserializeListMaybeAssignmentTargetAssignmentTargetWithDefault(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static ImmutableList<Maybe<AssignmentTargetAssignmentTargetWithDefault>> deserializeListMaybeAssignmentTargetAssignmentTargetWithDefault(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         JsonArray jsonArray = jsonElement.getAsJsonArray();
         if (jsonArray.size() == 0) {
           return ImmutableList.nil();
@@ -202,14 +192,14 @@ public class Deserializer {
         }
     }
 
-    private Maybe<AssignmentTarget> deserializeMaybeAssignmentTarget(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static Maybe<AssignmentTarget> deserializeMaybeAssignmentTarget(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         if (jsonElement.isJsonNull()) {
             return Maybe.nothing();
         }
         return Maybe.just((AssignmentTarget) deserializeNode(jsonElement));
     }
 
-    private ImmutableList<Maybe<BindingBindingWithDefault>> deserializeListMaybeBindingBindingWithDefault(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static ImmutableList<Maybe<BindingBindingWithDefault>> deserializeListMaybeBindingBindingWithDefault(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         JsonArray jsonArray = jsonElement.getAsJsonArray();
         if (jsonArray.size() == 0) {
           return ImmutableList.nil();
@@ -226,14 +216,14 @@ public class Deserializer {
         }
     }
 
-    private Maybe<Binding> deserializeMaybeBinding(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static Maybe<Binding> deserializeMaybeBinding(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         if (jsonElement.isJsonNull()) {
             return Maybe.nothing();
         }
         return Maybe.just((Binding) deserializeNode(jsonElement));
     }
 
-    private ImmutableList<Maybe<SpreadElementExpression>> deserializeListMaybeSpreadElementExpression(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static ImmutableList<Maybe<SpreadElementExpression>> deserializeListMaybeSpreadElementExpression(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         JsonArray jsonArray = jsonElement.getAsJsonArray();
         if (jsonArray.size() == 0) {
           return ImmutableList.nil();
@@ -250,14 +240,14 @@ public class Deserializer {
         }
     }
 
-    private Maybe<Expression> deserializeMaybeExpression(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static Maybe<Expression> deserializeMaybeExpression(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         if (jsonElement.isJsonNull()) {
             return Maybe.nothing();
         }
         return Maybe.just((Expression) deserializeNode(jsonElement));
     }
 
-    private ImmutableList<Statement> deserializeListStatement(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static ImmutableList<Statement> deserializeListStatement(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         JsonArray jsonArray = jsonElement.getAsJsonArray();
         if (jsonArray.size() == 0) {
           return ImmutableList.nil();
@@ -271,7 +261,7 @@ public class Deserializer {
         }
     }
 
-    private Maybe<String> deserializeMaybeString(JsonElement jsonElement) {
+    private static Maybe<String> deserializeMaybeString(JsonElement jsonElement) {
         if (jsonElement.isJsonNull()) {
             return Maybe.nothing();
         } else {
@@ -279,7 +269,7 @@ public class Deserializer {
         }
     }
 
-    private ImmutableList<SpreadElementExpression> deserializeListSpreadElementExpression(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static ImmutableList<SpreadElementExpression> deserializeListSpreadElementExpression(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         JsonArray jsonArray = jsonElement.getAsJsonArray();
         if (jsonArray.size() == 0) {
           return ImmutableList.nil();
@@ -293,7 +283,7 @@ public class Deserializer {
         }
     }
 
-    private ImmutableList<ClassElement> deserializeListClassElement(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static ImmutableList<ClassElement> deserializeListClassElement(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         JsonArray jsonArray = jsonElement.getAsJsonArray();
         if (jsonArray.size() == 0) {
           return ImmutableList.nil();
@@ -307,14 +297,14 @@ public class Deserializer {
         }
     }
 
-    private Maybe<BindingIdentifier> deserializeMaybeBindingIdentifier(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static Maybe<BindingIdentifier> deserializeMaybeBindingIdentifier(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         if (jsonElement.isJsonNull()) {
             return Maybe.nothing();
         }
         return Maybe.just((BindingIdentifier) deserializeNode(jsonElement));
     }
 
-    private ImmutableList<ExportFromSpecifier> deserializeListExportFromSpecifier(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static ImmutableList<ExportFromSpecifier> deserializeListExportFromSpecifier(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         JsonArray jsonArray = jsonElement.getAsJsonArray();
         if (jsonArray.size() == 0) {
           return ImmutableList.nil();
@@ -328,7 +318,7 @@ public class Deserializer {
         }
     }
 
-    private ImmutableList<ExportLocalSpecifier> deserializeListExportLocalSpecifier(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static ImmutableList<ExportLocalSpecifier> deserializeListExportLocalSpecifier(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         JsonArray jsonArray = jsonElement.getAsJsonArray();
         if (jsonArray.size() == 0) {
           return ImmutableList.nil();
@@ -342,14 +332,14 @@ public class Deserializer {
         }
     }
 
-    private Maybe<VariableDeclarationExpression> deserializeMaybeVariableDeclarationExpression(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static Maybe<VariableDeclarationExpression> deserializeMaybeVariableDeclarationExpression(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         if (jsonElement.isJsonNull()) {
             return Maybe.nothing();
         }
         return Maybe.just((VariableDeclarationExpression) deserializeNode(jsonElement));
     }
 
-    private ImmutableList<Parameter> deserializeListParameter(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static ImmutableList<Parameter> deserializeListParameter(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         JsonArray jsonArray = jsonElement.getAsJsonArray();
         if (jsonArray.size() == 0) {
           return ImmutableList.nil();
@@ -363,7 +353,7 @@ public class Deserializer {
         }
     }
 
-    private ImmutableList<Directive> deserializeListDirective(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static ImmutableList<Directive> deserializeListDirective(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         JsonArray jsonArray = jsonElement.getAsJsonArray();
         if (jsonArray.size() == 0) {
           return ImmutableList.nil();
@@ -377,14 +367,14 @@ public class Deserializer {
         }
     }
 
-    private Maybe<Statement> deserializeMaybeStatement(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static Maybe<Statement> deserializeMaybeStatement(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         if (jsonElement.isJsonNull()) {
             return Maybe.nothing();
         }
         return Maybe.just((Statement) deserializeNode(jsonElement));
     }
 
-    private ImmutableList<ImportSpecifier> deserializeListImportSpecifier(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static ImmutableList<ImportSpecifier> deserializeListImportSpecifier(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         JsonArray jsonArray = jsonElement.getAsJsonArray();
         if (jsonArray.size() == 0) {
           return ImmutableList.nil();
@@ -398,7 +388,7 @@ public class Deserializer {
         }
     }
 
-    private ImmutableList<ImportDeclarationExportDeclarationStatement> deserializeListImportDeclarationExportDeclarationStatement(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static ImmutableList<ImportDeclarationExportDeclarationStatement> deserializeListImportDeclarationExportDeclarationStatement(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         JsonArray jsonArray = jsonElement.getAsJsonArray();
         if (jsonArray.size() == 0) {
           return ImmutableList.nil();
@@ -412,7 +402,7 @@ public class Deserializer {
         }
     }
 
-    private ImmutableList<AssignmentTargetProperty> deserializeListAssignmentTargetProperty(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static ImmutableList<AssignmentTargetProperty> deserializeListAssignmentTargetProperty(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         JsonArray jsonArray = jsonElement.getAsJsonArray();
         if (jsonArray.size() == 0) {
           return ImmutableList.nil();
@@ -426,7 +416,7 @@ public class Deserializer {
         }
     }
 
-    private ImmutableList<BindingProperty> deserializeListBindingProperty(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static ImmutableList<BindingProperty> deserializeListBindingProperty(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         JsonArray jsonArray = jsonElement.getAsJsonArray();
         if (jsonArray.size() == 0) {
           return ImmutableList.nil();
@@ -440,7 +430,7 @@ public class Deserializer {
         }
     }
 
-    private ImmutableList<ObjectProperty> deserializeListObjectProperty(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static ImmutableList<ObjectProperty> deserializeListObjectProperty(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         JsonArray jsonArray = jsonElement.getAsJsonArray();
         if (jsonArray.size() == 0) {
           return ImmutableList.nil();
@@ -454,7 +444,7 @@ public class Deserializer {
         }
     }
 
-    private ImmutableList<SwitchCase> deserializeListSwitchCase(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static ImmutableList<SwitchCase> deserializeListSwitchCase(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         JsonArray jsonArray = jsonElement.getAsJsonArray();
         if (jsonArray.size() == 0) {
           return ImmutableList.nil();
@@ -468,7 +458,7 @@ public class Deserializer {
         }
     }
 
-    private ImmutableList<ExpressionTemplateElement> deserializeListExpressionTemplateElement(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static ImmutableList<ExpressionTemplateElement> deserializeListExpressionTemplateElement(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         JsonArray jsonArray = jsonElement.getAsJsonArray();
         if (jsonArray.size() == 0) {
           return ImmutableList.nil();
@@ -482,14 +472,14 @@ public class Deserializer {
         }
     }
 
-    private Maybe<CatchClause> deserializeMaybeCatchClause(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static Maybe<CatchClause> deserializeMaybeCatchClause(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         if (jsonElement.isJsonNull()) {
             return Maybe.nothing();
         }
         return Maybe.just((CatchClause) deserializeNode(jsonElement));
     }
 
-    private ImmutableList<VariableDeclarator> deserializeListVariableDeclarator(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+    private static ImmutableList<VariableDeclarator> deserializeListVariableDeclarator(JsonElement jsonElement) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         JsonArray jsonArray = jsonElement.getAsJsonArray();
         if (jsonArray.size() == 0) {
           return ImmutableList.nil();
@@ -504,7 +494,7 @@ public class Deserializer {
     }
 
 
-    protected Node deserializeNode(JsonElement jsonElement) throws ClassNotFoundException, IllegalAccessException, InstantiationException, NoSuchMethodException {
+    private static Node deserializeNode(JsonElement jsonElement) throws ClassNotFoundException, IllegalAccessException, InstantiationException, NoSuchMethodException {
         if (jsonElement.isJsonObject()) {
             JsonObject jsonObject = jsonElement.getAsJsonObject();
             if (jsonObject.has("type")) {
