+++ generated-sources/main/java/com/shapesecurity/shift/es2017/serialization/Serializer.java
@@ -1,4 +1,4 @@
-// Generated by reducer.js
+// Generated by serializer.js
 /**
  * Copyright 2018 Shape Security, Inc.
  *
@@ -16,18 +16,17 @@
  */
 
 
-package com.shapesecurity.shift.es2017.serialization;
+package com.shapesecurity.shift.serialization;
 
 import com.shapesecurity.functional.data.ImmutableList;
 import com.shapesecurity.functional.data.Maybe;
 import com.shapesecurity.functional.data.NonEmptyImmutableList;
 import com.shapesecurity.shift.es2017.ast.*;
-import com.shapesecurity.shift.es2017.ast.Module;
 import com.shapesecurity.shift.es2017.ast.operators.BinaryOperator;
 import com.shapesecurity.shift.es2017.ast.operators.CompoundAssignmentOperator;
 import com.shapesecurity.shift.es2017.ast.operators.UnaryOperator;
 import com.shapesecurity.shift.es2017.ast.operators.UpdateOperator;
-import com.shapesecurity.shift.es2017.utils.Utils;
+import com.shapesecurity.shift.utils.Utils;
 import com.shapesecurity.shift.es2017.reducer.Director;
 import com.shapesecurity.shift.es2017.reducer.Reducer;
 
@@ -74,91 +73,91 @@ public class Serializer implements Reducer<StringBuilder> {
         @Nonnull
         JsonObjectBuilder add(@Nonnull String property, boolean value) {
             optionalComma();
-            this.text.append(Utils.escapeStringLiteral(property, '"', false)).append(":").append(value);
+            this.text.append(Utils.escapeStringLiteral(property)).append(":").append(value);
             return this;
         }
 
         @Nonnull
         JsonObjectBuilder add(@Nonnull String property, @Nonnull String value) {
             optionalComma();
-            this.text.append(Utils.escapeStringLiteral(property, '"', false)).append(":").append(Utils.escapeStringLiteral(value, '"', false));
+            this.text.append(Utils.escapeStringLiteral(property)).append(":").append(Utils.escapeStringLiteral(value));
             return this;
         }
 
         @Nonnull
         JsonObjectBuilder add(@Nonnull String property, @Nonnull Number value) {
             optionalComma();
-            this.text.append(Utils.escapeStringLiteral(property, '"', false)).append(":").append(value);
+            this.text.append(Utils.escapeStringLiteral(property)).append(":").append(value);
             return this;
         }
 
         @Nonnull
         JsonObjectBuilder add(@Nonnull String property, @Nonnull BinaryOperator value) {
             optionalComma();
-            this.text.append(Utils.escapeStringLiteral(property, '"', false)).append(":").append(Utils.escapeStringLiteral(value.getName(), '"', false));
+            this.text.append(Utils.escapeStringLiteral(property)).append(":").append(Utils.escapeStringLiteral(value.getName()));
             return this;
         }
 
         @Nonnull
         JsonObjectBuilder add(@Nonnull String property, @Nonnull CompoundAssignmentOperator value) {
             optionalComma();
-            this.text.append(Utils.escapeStringLiteral(property, '"', false)).append(":").append(Utils.escapeStringLiteral(value.getName(), '"', false));
+            this.text.append(Utils.escapeStringLiteral(property)).append(":").append(Utils.escapeStringLiteral(value.getName()));
             return this;
         }
 
         @Nonnull
         JsonObjectBuilder add(@Nonnull String property, @Nonnull UnaryOperator value) {
             optionalComma();
-            this.text.append(Utils.escapeStringLiteral(property, '"', false)).append(":").append(Utils.escapeStringLiteral(value.getName(), '"', false));
+            this.text.append(Utils.escapeStringLiteral(property)).append(":").append(Utils.escapeStringLiteral(value.getName()));
             return this;
         }
 
         @Nonnull
         JsonObjectBuilder add(@Nonnull String property, @Nonnull UpdateOperator value) {
             optionalComma();
-            this.text.append(Utils.escapeStringLiteral(property, '"', false)).append(":").append(Utils.escapeStringLiteral(value.getName(), '"', false));
+            this.text.append(Utils.escapeStringLiteral(property)).append(":").append(Utils.escapeStringLiteral(value.getName()));
             return this;
         }
 
         @Nonnull
         JsonObjectBuilder add(@Nonnull String property, @Nonnull VariableDeclarationKind value) {
             optionalComma();
-            this.text.append(Utils.escapeStringLiteral(property, '"', false)).append(":").append(Utils.escapeStringLiteral(value.name, '"', false));
+            this.text.append(Utils.escapeStringLiteral(property)).append(":").append(Utils.escapeStringLiteral(value.name));
             return this;
         }
 
         @Nonnull
         JsonObjectBuilder add(@Nonnull String property, @Nonnull StringBuilder value) {
             optionalComma();
-            this.text.append(Utils.escapeStringLiteral(property, '"', false)).append(":").append(value);
+            this.text.append(Utils.escapeStringLiteral(property)).append(":").append(value);
             return this;
         }
 
         @Nonnull
         JsonObjectBuilder addMaybeString(@Nonnull String property, @Nonnull Maybe<String> value) {
             optionalComma();
-            this.text.append(Utils.escapeStringLiteral(property, '"', false)).append(":").append(value.map(Utils::escapeStringLiteral).orJust("null"));
+            this.text.append(Utils.escapeStringLiteral(property)).append(":").append(value.map(Utils::escapeStringLiteral).orJust("null"));
             return this;
         }
 
         @Nonnull
         JsonObjectBuilder add(@Nonnull String property, @Nonnull Maybe<StringBuilder> value) {
             optionalComma();
-            this.text.append(Utils.escapeStringLiteral(property, '"', false)).append(":").append(o(value));
+            this.text.append(Utils.escapeStringLiteral(property)).append(":").append(o(value));
             return this;
         }
 
         @Nonnull
         JsonObjectBuilder add(@Nonnull String property, @Nonnull ImmutableList<StringBuilder> value) {
             optionalComma();
-            this.text.append(Utils.escapeStringLiteral(property, '"', false)).append(":").append(list(value));
+            this.text.append(Utils.escapeStringLiteral(property)).append(":").append(list(value));
             return this;
         }
 
         @Nonnull
         JsonObjectBuilder addListMaybe(@Nonnull String property, @Nonnull ImmutableList<Maybe<StringBuilder>> value) { // because type erasure
             optionalComma();
-            this.text.append(Utils.escapeStringLiteral(property, '"', false)).append(":").append(list(value.map(Serializer::o)));
+            this.text.append(Utils.escapeStringLiteral(property)).append(":").append(list(value.map(Serializer::o)));
             return this;
         }
